Move warnAboutDataLoss into writeToStore.ts.
Warning about data loss is rightfully a concern of the StoreWriter, not
Policies. Also, StoreWriter has easy access to policies.hasMergeFunction,
which was the main selling point for putting warnAboutDataLoss in the
policies.ts module.
